@@ -3,7 +3,7 @@
     "description": "Command line interface library for PHP",
     "type": "library",
     "keywords": [
-        "php7",
+        "php8",
         "command",
         "argv-parser",
         "cli",
@@ -37,10 +37,10 @@
         }
     },
     "require": {
-        "php": ">=7.0"
+        "php": ">=8.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^6.0"
+        "phpunit/phpunit": "^9.0"
     },
     "scripts": {
         "test": "phpunit",
