@@ -27,49 +27,34 @@ use Ahc\Cli\IO\Interactor;
 class Application
 {
     /** @var Command[] */
-    protected $commands = [];
+    protected array $commands = [];
 
     /** @var array Raw argv sent to parse() */
-    protected $argv = [];
+    protected array $argv = [];
 
     /** @var array Command aliases [alias => cmd] */
-    protected $aliases = [];
-
-    /** @var string */
-    protected $name;
-
-    /** @var string App version */
-    protected $version = '';
+    protected array $aliases = [];
 
     /** @var string Ascii art logo */
-    protected $logo = '';
+    protected string $logo = '';
 
-    protected $default = '__default__';
+    protected string $default = '__default__';
 
     /** @var Interactor */
-    protected $io;
+    protected Interactor $io;
 
     /** @var callable The callable to perform exit */
     protected $onExit;
 
-    public function __construct(string $name, string $version = '0.0.1', callable $onExit = null)
+    public function __construct(protected string $name, protected string $version = '0.0.1', callable $onExit = null)
     {
-        $this->name    = $name;
-        $this->version = $version;
-
-        // @codeCoverageIgnoreStart
-        $this->onExit = $onExit ?? function ($exitCode = 0) {
-            exit($exitCode);
-        };
-        // @codeCoverageIgnoreEnd
+        $this->onExit = $onExit ?? fn (int $exitCode = 0) => exit($exitCode);
 
         $this->command('__default__', 'Default command', '', true)->on([$this, 'showHelp'], 'help');
     }
 
     /**
      * Get the name.
-     *
-     * @return string
      */
     public function name(): string
     {
@@ -78,8 +63,6 @@ class Application
 
     /**
      * Get the version.
-     *
-     * @return string
      */
     public function version(): string
     {
@@ -102,8 +85,6 @@ class Application
 
     /**
      * Get the raw argv.
-     *
-     * @return array
      */
     public function argv(): array
     {
@@ -130,14 +111,6 @@ class Application
 
     /**
      * Add a command by its name desc alias etc.
-     *
-     * @param string $name
-     * @param string $desc
-     * @param string $alias
-     * @param bool   $allowUnknown
-     * @param bool   $default
-     *
-     * @return Command
      */
     public function command(
         string $name,
@@ -155,12 +128,6 @@ class Application
 
     /**
      * Add a prepred command.
-     *
-     * @param Command $command
-     * @param string  $alias
-     * @param bool    $default
-     *
-     * @return self
      */
     public function add(Command $command, string $alias = '', bool $default = false): self
     {
@@ -192,17 +159,13 @@ class Application
 
     /**
      * Gets matching command for given argv.
-     *
-     * @param array $argv
-     *
-     * @return Command
      */
     public function commandFor(array $argv): Command
     {
         $argv += [null, null, null];
 
         return
-             // cmd
+            // cmd
             $this->commands[$argv[1]]
             // cmd alias
             ?? $this->commands[$this->aliases[$argv[1]] ?? null]
@@ -262,12 +225,8 @@ class Application
 
     /**
      * Handle the request, invoke action and call exit handler.
-     *
-     * @param array $argv
-     *
-     * @return mixed
      */
-    public function handle(array $argv)
+    public function handle(array $argv): mixed
     {
         if (\count($argv) < 2) {
             return $this->showHelp();
@@ -288,10 +247,6 @@ class Application
 
     /**
      * Get aliases for given command.
-     *
-     * @param Command $command
-     *
-     * @return array
      */
     protected function aliasesFor(Command $command): array
     {
@@ -309,10 +264,8 @@ class Application
 
     /**
      * Show help of all commands.
-     *
-     * @return mixed
      */
-    public function showHelp()
+    public function showHelp(): mixed
     {
         $writer = $this->io()->writer();
         $header = "{$this->name}, version {$this->version}";
@@ -336,12 +289,8 @@ class Application
 
     /**
      * Invoke command action.
-     *
-     * @param Command $command
-     *
-     * @return mixed
      */
-    protected function doAction(Command $command)
+    protected function doAction(Command $command): mixed
     {
         if ($command->name() === '__default__') {
             return $this->notFound();
@@ -351,7 +300,7 @@ class Application
         $command->interact($this->io());
 
         if (!$command->action() && !\method_exists($command, 'execute')) {
-            return;
+            return null;
         }
 
         $params = [];
@@ -368,10 +317,8 @@ class Application
 
     /**
      * Command not found handler.
-     *
-     * @return mixed
      */
-    protected function notFound()
+    protected function notFound(): mixed
     {
         $available = \array_keys($this->commands() + $this->aliases);
         $this->outputHelper()->showCommandNotFound($this->argv[1], $available);
