@@ -28,11 +28,10 @@ use Ahc\Cli\Output\Writer;
  */
 class OutputHelper
 {
-    /** @var Writer */
-    protected $writer;
+    protected Writer $writer;
 
     /** @var int Max width of command name */
-    protected $maxCmdName;
+    protected int $maxCmdName = 0;
 
     public function __construct(Writer $writer = null)
     {
@@ -41,12 +40,8 @@ class OutputHelper
 
     /**
      * Print stack trace and error msg of an exception.
-     *
-     * @param \Throwable $e
-     *
-     * @return void
      */
-    public function printTrace(\Throwable $e)
+    public function printTrace(\Throwable $e): void
     {
         $eClass = \get_class($e);
 
@@ -79,7 +74,7 @@ class OutputHelper
         $this->writer->colors($traceStr);
     }
 
-    protected function stringifyArgs(array $args)
+    protected function stringifyArgs(array $args): string
     {
         $holder = [];
 
@@ -90,7 +85,7 @@ class OutputHelper
         return \implode(', ', $holder);
     }
 
-    protected function stringifyArg($arg)
+    protected function stringifyArg($arg): string
     {
         if (\is_scalar($arg)) {
             return \var_export($arg, true);
@@ -144,9 +139,7 @@ class OutputHelper
      */
     public function showCommandsHelp(array $commands, string $header = '', string $footer = ''): self
     {
-        $this->maxCmdName = $commands ? \max(\array_map(function (Command $cmd) {
-            return \strlen($cmd->name());
-        }, $commands)) : 0;
+        $this->maxCmdName = $commands ? \max(\array_map(fn (Command $cmd) => \strlen($cmd->name()), $commands)) : 0;
 
         $this->showHelp('Commands', $commands, $header, $footer);
 
@@ -155,15 +148,8 @@ class OutputHelper
 
     /**
      * Show help with headers and footers.
-     *
-     * @param string $for
-     * @param array  $items
-     * @param string $header
-     * @param string $footer
-     *
-     * @return void
      */
-    protected function showHelp(string $for, array $items, string $header = '', string $footer = '')
+    protected function showHelp(string $for, array $items, string $header = '', string $footer = ''): void
     {
         if ($header) {
             $this->writer->bold($header, true);
@@ -195,16 +181,12 @@ class OutputHelper
      * Show usage examples of a Command.
      *
      * It replaces $0 with actual command name and properly pads ` ## ` segments.
-     *
-     * @param string $usage Usage description.
-     *
-     * @return self
      */
     public function showUsage(string $usage): self
     {
         $usage = \str_replace('$0', $_SERVER['argv'][0] ?? '[cmd]', $usage);
 
-        if (\strpos($usage, ' ## ') === false) {
+        if (!\str_contains($usage, ' ## ')) {
             $this->writer->eol()->boldGreen('Usage Examples:', true)->colors($usage)->eol();
 
             return $this;
@@ -217,7 +199,7 @@ class OutputHelper
             }
         }
 
-        $maxlen = \max($lines) + 4;
+        $maxlen = ($lines ? \max($lines) : 0) + 4;
         $usage  = \preg_replace_callback('~ ## ~', function () use (&$lines, $maxlen) {
             return \str_pad('# ', $maxlen - \array_shift($lines), ' ', \STR_PAD_LEFT);
         }, $usage);
@@ -257,15 +239,9 @@ class OutputHelper
      */
     protected function sortItems(array $items, &$max = 0): array
     {
-        $max = \max(\array_map(function ($item) {
-            return \strlen($this->getName($item));
-        }, $items));
+        $max = \max(\array_map(fn ($item) => \strlen($this->getName($item)), $items));
 
-        \uasort($items, function ($a, $b) {
-            /* @var Parameter $b */
-            /* @var Parameter $a */
-            return $a->name() <=> $b->name();
-        });
+        \uasort($items, fn ($a, $b) => $a->name() <=> $b->name());
 
         return $items;
     }
@@ -290,12 +266,8 @@ class OutputHelper
 
     /**
      * Get parameter label for humans.
-     *
-     * @param Parameter $item
-     *
-     * @return string
      */
-    protected function label(Parameter $item)
+    protected function label(Parameter $item): string
     {
         $name = $item->name();
 
