@@ -31,35 +31,17 @@ class Command extends Parser
 {
     use InflectsString;
 
-    /** @var callable */
-    protected $_action;
+    protected $_action = null;
 
-    /** @var string */
-    protected $_version;
+    protected string $_version = '';
 
-    /** @var string */
-    protected $_name;
+    protected string $_usage = '';
 
-    /** @var string */
-    protected $_desc;
+    protected ?string $_alias = null;
 
-    /** @var string Usage examples */
-    protected $_usage;
+    private array $_events = [];
 
-    /** @var string Command alias */
-    protected $_alias;
-
-    /** @var App The cli app this command is bound to */
-    protected $_app;
-
-    /** @var callable[] Events for options */
-    private $_events = [];
-
-    /** @var bool Whether to allow unknown (not registered) options */
-    private $_allowUnknown = false;
-
-    /** @var bool If the last seen arg was variadic */
-    private $_argVariadic = false;
+    private bool $_argVariadic = false;
 
     /**
      * Constructor.
@@ -69,46 +51,33 @@ class Command extends Parser
      * @param bool   $allowUnknown
      * @param App    $app
      */
-    public function __construct(string $name, string $desc = '', bool $allowUnknown = false, App $app = null)
-    {
-        $this->_name         = $name;
-        $this->_desc         = $desc;
-        $this->_allowUnknown = $allowUnknown;
-        $this->_app          = $app;
-
+    public function __construct(
+        protected string $_name,
+        protected string $_desc = '',
+        protected bool $_allowUnknown = false,
+        protected ?App $_app = null
+    ) {
         $this->defaults();
     }
 
     /**
      * Sets default options, actions and exit handler.
-     *
-     * @return self
      */
     protected function defaults(): self
     {
         $this->option('-h, --help', 'Show help')->on([$this, 'showHelp']);
         $this->option('-V, --version', 'Show version')->on([$this, 'showVersion']);
-        $this->option('-v, --verbosity', 'Verbosity level', null, 0)->on(function () {
-            $this->set('verbosity', ($this->verbosity ?? 0) + 1);
+        $this->option('-v, --verbosity', 'Verbosity level', null, 0)->on(
+            fn () => $this->set('verbosity', ($this->verbosity ?? 0) + 1) && false
+        );
 
-            return false;
-        });
-
-        // @codeCoverageIgnoreStart
-        $this->onExit(function ($exitCode = 0) {
-            exit($exitCode);
-        });
-        // @codeCoverageIgnoreEnd
+        $this->onExit(fn ($exitCode = 0) => exit($exitCode));
 
         return $this;
     }
 
     /**
      * Sets version.
-     *
-     * @param string $version
-     *
-     * @return self
      */
     public function version(string $version): self
     {
@@ -119,8 +88,6 @@ class Command extends Parser
 
     /**
      * Gets command name.
-     *
-     * @return string
      */
     public function name(): string
     {
@@ -129,8 +96,6 @@ class Command extends Parser
 
     /**
      * Gets command description.
-     *
-     * @return string
      */
     public function desc(): string
     {
@@ -139,20 +104,14 @@ class Command extends Parser
 
     /**
      * Get the app this command belongs to.
-     *
-     * @return null|App
      */
-    public function app()
+    public function app(): ?App
     {
         return $this->_app;
     }
 
     /**
      * Bind command to the app.
-     *
-     * @param App|null $app
-     *
-     * @return self
      */
     public function bind(App $app = null): self
     {
@@ -163,10 +122,6 @@ class Command extends Parser
 
     /**
      * Registers argument definitions (all at once). Only last one can be variadic.
-     *
-     * @param string $definitions
-     *
-     * @return self
      */
     public function arguments(string $definitions): self
     {
@@ -181,12 +136,6 @@ class Command extends Parser
 
     /**
      * Register an argument.
-     *
-     * @param string $raw
-     * @param string $desc
-     * @param mixed  $default
-     *
-     * @return self
      */
     public function argument(string $raw, string $desc = '', $default = null): self
     {
@@ -207,13 +156,6 @@ class Command extends Parser
 
     /**
      * Registers new option.
-     *
-     * @param string        $raw
-     * @param string        $desc
-     * @param callable|null $filter
-     * @param mixed         $default
-     *
-     * @return self
      */
     public function option(string $raw, string $desc = '', callable $filter = null, $default = null): self
     {
@@ -226,8 +168,6 @@ class Command extends Parser
 
     /**
      * Gets user options (i.e without defaults).
-     *
-     * @return array
      */
     public function userOptions(): array
     {
@@ -276,11 +216,6 @@ class Command extends Parser
 
     /**
      * Sets event handler for last (or given) option.
-     *
-     * @param callable $fn
-     * @param string   $option
-     *
-     * @return self
      */
     public function on(callable $fn, string $option = null): self
     {
@@ -293,10 +228,6 @@ class Command extends Parser
 
     /**
      * Register exit handler.
-     *
-     * @param callable $fn
-     *
-     * @return self
      */
     public function onExit(callable $fn): self
     {
@@ -308,7 +239,7 @@ class Command extends Parser
     /**
      * {@inheritdoc}
      */
-    protected function handleUnknown(string $arg, string $value = null)
+    protected function handleUnknown(string $arg, string $value = null): mixed
     {
         if ($this->_allowUnknown) {
             return $this->set($this->toCamelCase($arg), $value);
@@ -329,10 +260,8 @@ class Command extends Parser
 
     /**
      * Shows command help then aborts.
-     *
-     * @return mixed
      */
-    public function showHelp()
+    public function showHelp(): mixed
     {
         $io     = $this->io();
         $helper = new OutputHelper($io->writer());
@@ -354,10 +283,8 @@ class Command extends Parser
 
     /**
      * Shows command version then aborts.
-     *
-     * @return mixed
      */
-    public function showVersion()
+    public function showVersion(): mixed
     {
         $this->writer()->bold($this->_version, true);
 
@@ -367,7 +294,7 @@ class Command extends Parser
     /**
      * {@inheritdoc}
      */
-    public function emit(string $event, $value = null)
+    public function emit(string $event, $value = null): mixed
     {
         if (empty($this->_events[$event])) {
             return null;
@@ -378,24 +305,16 @@ class Command extends Parser
 
     /**
      * Tap return given object or if that is null then app instance. This aids for chaining.
-     *
-     * @param mixed $object
-     *
-     * @return mixed
      */
-    public function tap($object = null)
+    public function tap(object $object = null)
     {
         return $object ?? $this->_app;
     }
 
     /**
      * Performs user interaction if required to set some missing values.
-     *
-     * @param Interactor $io
-     *
-     * @return void
      */
-    public function interact(Interactor $io)
+    public function interact(Interactor $io): void
     {
         // Subclasses will do the needful.
     }
@@ -420,8 +339,6 @@ class Command extends Parser
 
     /**
      * Get a writer instance.
-     *
-     * @return Writer
      */
     protected function writer(): Writer
     {
@@ -430,8 +347,6 @@ class Command extends Parser
 
     /**
      * Get IO instance.
-     *
-     * @return Interactor
      */
     protected function io(): Interactor
     {
