@@ -21,16 +21,14 @@ namespace Ahc\Cli\Input;
  */
 class Option extends Parameter
 {
-    /** @var string Short name */
-    protected $short = '';
+    protected string $short = '';
 
-    /** @var string Long name */
-    protected $long = '';
+    protected string $long = '';
 
     /**
      * {@inheritdoc}
      */
-    protected function parse(string $raw)
+    protected function parse(string $raw): void
     {
         if (\strpos($raw, '-with-') !== false) {
             $this->default = false;
@@ -50,8 +48,6 @@ class Option extends Parameter
 
     /**
      * Get long name.
-     *
-     * @return string
      */
     public function long(): string
     {
@@ -60,8 +56,6 @@ class Option extends Parameter
 
     /**
      * Get short name.
-     *
-     * @return string
      */
     public function short(): string
     {
@@ -70,10 +64,6 @@ class Option extends Parameter
 
     /**
      * Test if this option matches given arg.
-     *
-     * @param string $arg
-     *
-     * @return bool
      */
     public function is(string $arg): bool
     {
@@ -82,8 +72,6 @@ class Option extends Parameter
 
     /**
      * Check if the option is boolean type.
-     *
-     * @return bool
      */
     public function bool(): bool
     {
