@@ -21,7 +21,7 @@ class WriterTest extends CliTestCase
     {
         (new Writer)->write('Hey');
 
-        $this->assertContains('Hey', $this->buffer());
+        $this->assertStringContainsString('Hey', $this->buffer());
         $this->assertSame("\033[0;37mHey\033[0m", $this->buffer());
     }
 
@@ -29,7 +29,7 @@ class WriterTest extends CliTestCase
     {
         (new Writer)->error->write('Something wrong');
 
-        $this->assertContains('Something wrong', $this->buffer());
+        $this->assertStringContainsString('Something wrong', $this->buffer());
         $this->assertSame("\033[0;31mSomething wrong\033[0m", $this->buffer());
     }
 
@@ -37,7 +37,7 @@ class WriterTest extends CliTestCase
     {
         (new Writer)->write('Hello', true);
 
-        $this->assertContains('Hello', $this->buffer());
+        $this->assertStringContainsString('Hello', $this->buffer());
         $this->assertSame("\033[0;37mHello\033[0m" . PHP_EOL, $this->buffer());
     }
 
@@ -45,7 +45,7 @@ class WriterTest extends CliTestCase
     {
         (new Writer)->bold->red->bgGreen->write('bold->red->bgGreen');
 
-        $this->assertContains('bold->red->bgGreen', $this->buffer());
+        $this->assertStringContainsString('bold->red->bgGreen', $this->buffer());
         $this->assertSame("\033[1;31;42mbold->red->bgGreen\033[0m", $this->buffer());
     }
 
